requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
comm==0.2.1
decorator==5.1.1
Django==5.0.1
django-ckeditor==6.7.0
django-js-asset==2.2.0
executing==2.0.1
ipython==8.20.0
ipywidgets==8.1.1
jedi==0.19.1
jupyterlab-widgets==3.0.9
matplotlib-inline==0.1.6
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2023.4
wcwidth==0.2.13
widgetsnbextension==4.0.9
